Move junit property fixtures and reporting hook
This was in conftest, which was getting copied for the unit test.

Since the file with the hooks/fixtures needs to be copied during the
unit test execution, split the reporting fixtures into their own file
for more specific unit testing.

This allows for content in foreman/conftest.py to not have to be unit
test safe, when the module is not intended for use in unit tests
